pyproject.toml
[project]
name = "ml-finance-agent"
version = "0.1.0"
description = "AI-powered agent for credit risk analysis"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3",
    "scikit-learn>=1.8",
    "mlflow>=3.10",
    "torch>=2.11",
    "langchain>=1.2",
    "langchain-openai>=1.1",
    "langgraph>=1.1",
    "faiss-cpu>=1.13",
    "sentence-transformers>=5.4",
    "ragas>=0.4",
    "datasets>=4.8",
    "python-dotenv>=1.1",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "joblib>=1.5",
    "jupyter>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "ruff>=0.11",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]
omit = ["src/models/train.py", "src/agent/*", "src/serving/*"]
